Вынес логику проверки хэштегов, подправил логику работы кнопок увеличения масштаба фото, плюс маленькие доработки
