+++ js/main.js
@@ -1,5 +1,4 @@
 import { generateUserInterface } from './miniature.js';
 import {} from './modal.js';
-import {} from './form.js';
-import {} from './filter.js';
+import {} from './upload-form.js';
 generateUserInterface();
